Move FileSystemExtensionContainer from "Internal" to "FileSystem" namespace
