+++ Source/Testably.Abstractions/FileSystem/FileSystemInfoWrapper.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using Testably.Abstractions.Internal;
 
 namespace Testably.Abstractions.FileSystem;
